requirements.txt
xgboost==1.1.1
seaborn==0.10.1
scipy==1.5.0
scikit-learn==0.23.2
pandas==1.1.0
numpy==1.19.1
matplotlib==3.3.1
lightgbm==2.3.0
joblib==0.16.0
flasgger==0.9.5
flask-request-validator==2.1.2
feature-engine==0.6.1
